iqss/dataverse#3031 - stage set, WorldMapLayerInfo prepped for shapefiles
